server/src/main/java/ru/practicum/shareit/request/RequestValidation.java
package ru.practicum.shareit.request;

import lombok.AllArgsConstructor;
import org.springframework.stereotype.Service;
import ru.practicum.shareit.exception.ValidationException;
import ru.practicum.shareit.request.dto.ItemRequestDto;

@Service
@AllArgsConstructor
public class RequestValidation {

    public void checkValidateResponse(ItemRequestDto requestDto) {
        if (requestDto.getDescription() == null || requestDto.getDescription().isBlank()) {
            throw new ValidationException("Запрос не может быть пустым");
        }
    }
}
